Dev joao

+++ backend/src/main/java/pt/ulisboa/tecnico/cmov/shopist/pojo/Coordinates.java
@@ -5,8 +5,8 @@ import lombok.Data;
 
 @Data
 public class Coordinates {
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
     @Override
     public boolean equals(Object o) {
